requirements.txt
fastapi
uvicorn
Pillow
python-multipart
pydantic
typing-extensions
psutil
requests
python-dotenv
websockets
python-socketio
numpy